Added new problem

+++ Codeforces/theatre-square-1-a.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main()
 {
-    int n,m,a;
+    long long n,m,a;
     cin>>n>>m>>a;
     long long ans=((n/a)+(n%a!=0))*(m/a+(m%a!=0));
     cout<<ans;
